pyproject.toml
[project]
name = "lulzcasz"
version = "0.15.2"
description = "lulzcasz.dev source code"
readme = "README.md"
requires-python = "~=3.12.0"
dependencies = [
    "celery~=5.5",
    "django~=5.2",
    "django-modeltranslation~=0.19",
    "django-polymorphic~=4.1",
    "django-storages[s3]~=1.14",
    "django-taggit~=6.1",
    "django-tinymce~=5.0",
    "django-treebeard~=4.7",
    "gunicorn~=23.0",
    "pillow~=12.0",
    "psycopg~=3.2",
    "redis~=7.0",
    "whitenoise[brotli]~=6.11",
]

[dependency-groups]
dev = [
    "ipython~=9.7",
    "psycopg[binary]~=3.2",
]
